relocate mode kernel code
