+++ src/ATen/native/xpu/sycl/TensorCompareKernels.h
@@ -27,11 +27,4 @@ TORCH_XPU_API void isin_kernel(
     bool invert,
     const Tensor& out);
 
-TORCH_XPU_API void mode_kernel(
-    Tensor& values,
-    Tensor& indices,
-    const Tensor& self,
-    int64_t dim,
-    bool keepdim);
-
 } // namespace at::native::xpu
